Uploaded the Toy Training Set creation script
We are already pretty low on time. A hacky, yet somehow efficient method is to split 20% off the training data and call it a 'toy training set'. We run all our experiments on this train set, and given good performance, shortlist the model for training on the entire training set.